Topsis-Chirag-102303003/topsis-web-service/requirements.txt
fastapi
uvicorn
python-multipart
pandas
numpy
python-dotenv
email-validator
